Moved setting format validations to concern
