+++ app/models/setting.rb
@@ -1,13 +1,9 @@
 class Setting < ApplicationRecord
+  include Concerns::FormatValidator
   validates :code, presence: true
   validates :description, presence: true
   validates :value, presence: true
   validates :code, uniqueness: true
-  validate :validate_value
-
-  FORMAT_VALIDATIONS = {
-    terms_and_conditions_link: [:validate_json_hash],
-  }.with_indifferent_access.freeze
 
   def self.auction_currency
     Setting.find_by(code: :auction_currency).value
@@ -112,28 +108,5 @@ class Setting < ApplicationRecord
     rescue NoMethodError
       false
     end
-  end
-
-  private
-
-  def validate_value
-    methods = FORMAT_VALIDATIONS[code]
-    return unless methods
-
-    methods.each do |method|
-      send(method)
-    end
-  end
-
-  def validate_json_hash
-    return if parsable_json_hash?(value)
-
-    errors.add(:value, :invalid)
-  end
-
-  def parsable_json_hash?(value)
-    JSON.parse(value).is_a?(Hash)
-  rescue JSON::ParserError
-    errors.add(:value, :not_a_json_hash)
   end
 end
