+++ lib/tasks/data_migrations/convert_terms_and_conditions_link_setting.rake
@@ -7,9 +7,9 @@ namespace :data_migrations do
         Link to terms and conditions document. Must be single parsable hash of <locale>:<URL> elements.
         URL can be relative ('/public/terms_and_conditions.pdf') 
         or absolute ('https://example.com/terms_and_conditions.pdf'). Relative URL must start with a
-        forward slash. Default is: "{\"en\":\"https://example.com\"}"
+        forward slash. Default is: "{\"en\":\"https://example.com\", \"et\":\"https://example.et\"}"
       TEXT
-      value = "{\"en\":\"https://example.com\"}"
+      value = "{\"en\":\"https://example.com\", \"et\":\"https://example.et\"}"
       setting.update(value: value,
                      description: terms_and_conditions_description)
     end
